+++ sdk/azcore/runtime/policy_logging.go
@@ -15,10 +15,10 @@ import (
 	"strings"
 	"time"
 
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/internal/log"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/internal/shared"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/internal/diag"
-	"github.com/Azure/azure-sdk-for-go/sdk/internal/log"
 )
 
 type logPolicy struct {